chore(query): relocate query-adjacent modules from posthog/caching
