+++ posthog/hogql_queries/insights/lifecycle/lifecycle_query_runner.py
@@ -26,7 +26,6 @@ from posthog.hogql.printer import to_printed_hogql
 from posthog.hogql.property import action_to_expr, property_to_expr
 from posthog.hogql.query import execute_hogql_query
 
-from posthog.caching.insights_api import BASE_MINIMUM_INSIGHT_REFRESH_INTERVAL, REDUCED_MINIMUM_INSIGHT_REFRESH_INTERVAL
 from posthog.clickhouse.query_tagging import tag_contains_user_hogql
 from posthog.hogql_queries.insights.lifecycle.lifecycle_validation_rules import (
     RequireLifecycleDataWarehouseSeriesForCustomAggregationTarget,
@@ -39,6 +38,10 @@ from posthog.hogql_queries.validation.validation import QueryValidationRule
 from posthog.models.filters.mixins.utils import cached_property
 
 from products.actions.backend.models.action import Action
+from products.product_analytics.backend.insights_api import (
+    BASE_MINIMUM_INSIGHT_REFRESH_INTERVAL,
+    REDUCED_MINIMUM_INSIGHT_REFRESH_INTERVAL,
+)
 
 
 class LifecycleQueryRunner(AnalyticsQueryRunner[LifecycleQueryResponse]):
