+++ posthog/hogql_queries/test/test_hogql_query_runner.py
@@ -20,8 +20,8 @@ from posthog.hogql.errors import ExposedHogQLError, QueryError
 from posthog.hogql.user_query_validator import HOGQL_PERSONAL_API_KEY_OFFSET_ALLOWED_FLAG, OFFSET_NOT_ALLOWED_MESSAGE
 from posthog.hogql.visitor import clear_locations
 
-from posthog.caching.utils import ThresholdMode, staleness_threshold_map
 from posthog.hogql_queries.hogql_query_runner import HogQLQueryRunner
+from posthog.hogql_queries.staleness import ThresholdMode, staleness_threshold_map
 from posthog.models.utils import UUIDT
 
 from products.product_analytics.backend.models.insight_variable import InsightVariable
