+++ posthog/query_cache/backend.py
@@ -1,0 +1,2 @@
+# The Django CACHES alias whose Redis holds every query cache outcome.
+QUERY_CACHE_ALIAS = "query_cache"
