+++ posthog/tasks/alerts/test/test_trends_absolute_alerts.py
@@ -24,11 +24,11 @@ from posthog.schema import (
 )
 
 from posthog.api.test.dashboards import DashboardAPI
-from posthog.caching.calculate_results import calculate_for_query_based_insight
 from posthog.models.instance_setting import set_instance_setting
 from posthog.tasks.alerts.test.alert_check_helpers import run_alert_check
 
 from products.alerts.backend.models.alert import AlertCheck, AlertConfiguration
+from products.product_analytics.backend.calculate_results import calculate_for_query_based_insight
 
 # 8:55 AM
 FROZEN_TIME = dateutil.parser.parse("2024-06-02T08:55:00.000Z")
