+++ posthog/tasks/alerts/trends.py
@@ -4,9 +4,10 @@ import numpy as np
 
 from posthog.schema import FunnelsQuery, IntervalType, TrendsAlertConfig, TrendsQuery
 
-from posthog.caching.insight_result import InsightResult
 from posthog.hogql_queries.insights.utils.breakdowns import has_breakdown_filter
 from posthog.tasks.alerts.utils import NON_TIME_SERIES_DISPLAY_TYPES
+
+from products.product_analytics.backend.insight_result import InsightResult
 
 
 # TODO: move the TrendResult UI type to schema.ts and use that instead
