+++ posthog/temporal/tests/ai/test_anomaly_investigation_tools.py
@@ -4,10 +4,10 @@ from unittest.mock import MagicMock, patch
 
 from posthog.schema import BaseMathType, ChartDisplayType, EventsNode, IntervalType, TrendsFilter, TrendsQuery
 
-from posthog.caching.insight_result import InsightResult
 from posthog.temporal.ai.anomaly_investigation.tools import _run_detector_simulation
 
 from products.alerts.backend.models.alert import AlertConfiguration
+from products.product_analytics.backend.insight_result import InsightResult
 from products.product_analytics.backend.models.insight import Insight
