+++ products/alerts/backend/evaluation/test/test_high_frequency_freshness.py
@@ -18,7 +18,6 @@ from posthog.schema import (
 )
 
 from posthog.api.services.query import ExecutionMode
-from posthog.caching.insight_result import InsightResult
 
 from products.alerts.backend.evaluation.contract import execution_mode_for_alert
 from products.alerts.backend.evaluation.detector import extract_detector_series
@@ -27,6 +26,7 @@ from products.alerts.backend.evaluation.funnels import FunnelsExtractor
 from products.alerts.backend.evaluation.hogql import HogQLExtractor
 from products.alerts.backend.evaluation.trends import TrendsExtractor
 from products.alerts.backend.models.alert import AlertConfiguration
+from products.product_analytics.backend.insight_result import InsightResult
 from products.product_analytics.backend.models.insight import Insight
 
 ALWAYS = ExecutionMode.CALCULATE_BLOCKING_ALWAYS
