+++ products/product_analytics/backend/test/test_warming.py
@@ -3,12 +3,16 @@ from datetime import UTC, datetime, timedelta
 from posthog.test.base import APIBaseTest
 from unittest.mock import patch
 
-from posthog.caching.warming import insights_to_keep_fresh, schedule_warming_for_teams_task, warm_insight_cache_task
 from posthog.exceptions import ClickHouseAtCapacity
 
 from products.dashboards.backend.models.dashboard import Dashboard
 from products.dashboards.backend.models.dashboard_tile import DashboardTile
 from products.product_analytics.backend.models.insight import Insight, InsightViewed
+from products.product_analytics.backend.warming import (
+    insights_to_keep_fresh,
+    schedule_warming_for_teams_task,
+    warm_insight_cache_task,
+)
 
 
 class TestWarming(APIBaseTest):
@@ -49,13 +53,13 @@ class TestWarming(APIBaseTest):
             team=self.team, user=self.user, insight=self.insight5, last_viewed_at=datetime.now(UTC) - timedelta(days=1)
         )
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_no_stale_insights(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = []
         insights = list(insights_to_keep_fresh(self.team))
         self.assertEqual(insights, [])
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_no_stale_dashboard_insights(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = [
             "2345:",
@@ -66,7 +70,7 @@ class TestWarming(APIBaseTest):
         ]
         self.assertEqual(insights, exptected_results)
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_only_insights_with_dashboards(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = [
             "1234:5678",
@@ -78,7 +82,7 @@ class TestWarming(APIBaseTest):
         ]
         self.assertEqual(insights, expected_results)
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_mixed_valid_and_invalid_combos(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = [
             "1234:5678",
@@ -92,19 +96,19 @@ class TestWarming(APIBaseTest):
         ]
         self.assertEqual(insights, expected_results)
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_insights_not_viewed_recently(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = ["4567:"]
         insights = list(insights_to_keep_fresh(self.team))
         self.assertEqual(insights, [])
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_dashboards_not_accessed_recently(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = ["5678:8901"]
         insights = list(insights_to_keep_fresh(self.team))
         self.assertEqual(insights, [])
 
-    @patch("posthog.caching.warming.get_stale_insights")
+    @patch("products.product_analytics.backend.warming.get_stale_insights")
     def test_insights_to_keep_fresh_combination_of_cases(self, mock_get_stale_insights):
         mock_get_stale_insights.return_value = [
             "1234:5678",
@@ -127,9 +131,9 @@ class TestScheduleWarmingForTeamsTask(APIBaseTest):
         self.team1 = self.create_team_with_organization(organization=self.organization)
         self.team2 = self.create_team_with_organization(organization=self.organization)
 
-    @patch("posthog.caching.warming.largest_teams")
-    @patch("posthog.caching.warming.insights_to_keep_fresh")
-    @patch("posthog.caching.warming.warm_insight_cache_task.si")
+    @patch("products.product_analytics.backend.warming.largest_teams")
+    @patch("products.product_analytics.backend.warming.insights_to_keep_fresh")
+    @patch("products.product_analytics.backend.warming.warm_insight_cache_task.si")
     def test_schedule_warming_for_teams_task_with_empty_insight_tuples(
         self, mock_warm_insight_cache_task_si, mock_insights_to_keep_fresh, mock_largest_teams
     ):
@@ -141,9 +145,9 @@ class TestScheduleWarmingForTeamsTask(APIBaseTest):
         mock_insights_to_keep_fresh.assert_called()
         mock_warm_insight_cache_task_si.assert_not_called()
 
-    @patch("posthog.caching.warming.largest_teams")
-    @patch("posthog.caching.warming.insights_to_keep_fresh")
-    @patch("posthog.caching.warming.warm_insight_cache_task.si")
+    @patch("products.product_analytics.backend.warming.largest_teams")
+    @patch("products.product_analytics.backend.warming.insights_to_keep_fresh")
+    @patch("products.product_analytics.backend.warming.warm_insight_cache_task.si")
     def test_schedule_warming_for_teams_task_with_non_empty_insight_tuples(
         self, mock_warm_insight_cache_task_si, mock_insights_to_keep_fresh, mock_largest_teams
     ):
@@ -160,8 +164,8 @@ class TestScheduleWarmingForTeamsTask(APIBaseTest):
 
 
 class TestWarmInsightCacheTask(APIBaseTest):
-    @patch("posthog.caching.warming.capture_exception")
-    @patch("posthog.caching.warming.process_query_dict", side_effect=ClickHouseAtCapacity())
+    @patch("products.product_analytics.backend.warming.capture_exception")
+    @patch("products.product_analytics.backend.warming.process_query_dict", side_effect=ClickHouseAtCapacity())
     def test_capacity_errors_propagate_for_retry_instead_of_being_captured(
         self, mock_process_query_dict, mock_capture_exception
     ):
