+++ products/pulse/backend/sources/strategy.py
@@ -5,11 +5,11 @@ from django.db.models import QuerySet
 import structlog
 
 from posthog.api.services.query import ExecutionMode
-from posthog.caching.calculate_results import calculate_for_query_based_insight
 from posthog.exceptions_capture import capture_exception
 from posthog.models.team import Team
 from posthog.rbac.user_access_control import UserAccessControl
 
+from products.product_analytics.backend.calculate_results import calculate_for_query_based_insight
 from products.product_analytics.backend.models.insight import Insight
 from products.pulse.backend.config import BriefSettings
 from products.pulse.backend.sources.base import EvidenceRef, EvidenceType, SourceItem, SourceItemKind
